third_party/blink/web_tests/fast/css/invalidation/detach-reattach-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS getComputedStyle(inner).backgroundColor is "rgba(0, 0, 0, 0)"
